requirements.txt
flask
sqlalchemy
bcrypt
psycopg2-binary
flask-jwt-extended
pymongo
pandas
xlrd
python-dotenv